veh_aero/requirements.txt
streamlit>=1.28.0
numpy==1.26.4
opencv-python>=4.8.0
matplotlib>=3.8.0
boto3>=1.33.0
Pillow>=10.0.0
setuptools>=69.0.0
